# Changed locking strategy