--- a/bigcache.go
+++ b/bigcache.go
@@ -82,6 +82,7 @@ func (c *BigCache) Get(key string) ([]byte, error) {
 		if c.config.Verbose {
 			log.Printf("Collision detected. Both %q and %q have the same hash %x", key, entryKey, hashedKey)
 		}
+		shard.lock.RUnlock()
 		return nil, notFound(key)
 	}
 	return readEntry(wrappedEntry), nil
@@ -92,7 +93,6 @@ func (c *BigCache) Set(key string, entry []byte) error {
 	hashedKey := c.hash.Sum64(key)
 	shard := c.getShard(hashedKey)
 	shard.lock.Lock()
-	defer shard.lock.Unlock()
 
 	currentTimestamp := uint64(c.clock.epoch())
 
@@ -111,8 +111,10 @@ func (c *BigCache) Set(key string, entry []byte) error {
 	for {
 		if index, err := shard.entries.Push(w); err == nil {
 			shard.hashmap[hashedKey] = uint32(index)
+			shard.lock.Unlock()
 			return nil
 		} else if shard.removeOldestEntry() != nil {
+			shard.lock.Unlock()
 			return fmt.Errorf("Entry is bigger than max shard size.")
 		}
 	}