There is a known issue in Go with deferred operations being much slower than explicit operations, both directly and indirectly impacting locking.

Here are the performance comparisons (from benchstat) over 1000 iterations:

```
name                   old time/op    new time/op    delta
BigCacheSet-8             569ns ± 6%     511ns ± 5%  -10.24%  (p=0.000 n=860+949)
BigCacheGet-8             525ns ± 3%     454ns ± 5%  -13.45%  (p=0.000 n=922+930)
BigCacheSetParallel-8     178ns ± 2%     164ns ±20%   -7.61%  (p=0.000 n=961+804)
```

I did this over 1000 iterations because benchmarking nanosecond operations is hard to get consistent, so that gives a good view of the larger performance gain.

Generally, on OS X and Windows, the change gives between 10-15% performance gain.

Signed-off-by: Mike Lloyd <mlloyd@pivotal.io>