@janisz 2017-04-15
Tests hangs on Travis. Any idea why?