@mxplusb 2017-04-15
No, it was hanging on my computer too. I didn't make any changes to the
tests, just the locking statements. Might be a deadlock, I can look into
it.
On Fri, Apr 14, 2017 at 10:55 PM janisz <notifications@github.com> wrote:

> Tests hangs on Travis. Any idea why?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/allegro/bigcache/pull/37#issuecomment-294274513>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AHMEsRD7xsl07IqeKnCYo28KuquMrcksks5rwFvQgaJpZM4M-MRu>
> .
>
-- 
Thanks!

Mike Lloyd
USMC
808-633-8998