@janisz 2017-04-15
In get method not all paths unlock mutex. Can you refactor code to have only one return statement followed by unlock? This will make it easier to see the flow, lock at the beginning and unlock just after return.